Resolve a todo.
I planned to remove the subsection, but I figure we need not to for the
following reasons. (1) It is nice for people to see where skein
categories are defined in the table of content. (2) This subsection
isn't really that short, and we may be adding more things into it
later. (3) The subsection for pre-skein category is a bit long. Merging
this subsection into it may weaken the focus of it.